backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
chromadb>=0.4.0
ollama>=0.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.0.0
lxml>=5.0.0
sse-starlette>=1.6.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
websockets>=12.0
pgvector>=0.3.0
